Hide highlight when owner moves
I couldn't solve handling stage movement and resizing in a reliable way across platforms. Attempts resulted in windows behaving in a jerky and unreliably way.

So this commit causes the highlight to simply be removed whenever the owner window is moved or resized, in the hope that 'gone' is better than 'wrong'.