Alternative node highlighting using an Effect
Alternative tour node highlighting by attaching a JavaFX `Effect` (as long as the node does not already have an effect).

The main advantage of this over CSS is that it supports adding both an inner and outer shadow.

It would also potentially be easier to change the color programmatically... but CSS customizations will be ignored.